feat: reorg to multiple crates
Add top level cargo toml since it helps with IDE support.
