+++ Cargo.toml
@@ -1,0 +1,4 @@
+[workspace]
+members = [
+  "api/rust"
+]
