+++ api/rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "suborbital"
+name = "suborbital-core"
 version = "0.12.0"
 authors = ["cohix <connorjhicks@gmail.com>", "Florian Fromm <flrn.frmm@gmail.com>"]
 edition = "2018"
@@ -12,4 +12,5 @@ homepage = "https://suborbital.dev"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-codegen = { path = "./codegen" }
+suborbital-core = {path = "crates/core"}
+codegen = {path = "crates/codegen"}
