+++ api/rust/crates/core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "suborbital-core"
+version = "0.12.0"
+authors = ["cohix <connorjhicks@gmail.com>", "Florian Fromm <flrn.frmm@gmail.com>"]
+edition = "2018"
+description = "Suborbital Wasm Runnable API"
+license = "Apache-2.0"
+repository = "https://github.com/suborbital/reactr"
+homepage = "https://suborbital.dev"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+codegen = { path = "../codegen" }
