+++ api/rust/crates/core/src/util.rs
