+++ rustfmt.toml
@@ -1,0 +1,1 @@
+hard_tab=true
